pyproject.toml
[project]
name = "mlaude"
version = "0.3.0-dev"
description = "Local AI agent — CLI-first, tool-calling, self-improving"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "rich>=13.0.0",
    "typer[all]>=0.9.0",
    "prompt_toolkit>=3.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.4.0",
    "pexpect>=4.9.0",
]
telegram = [
    "python-telegram-bot>=21.0",
]
openai = [
    "openai>=1.50.0",
]
anthropic = [
    "anthropic>=0.40.0",
]
gateway = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
browser = [
    "playwright>=1.52.0",
]
search = [
    "readability-lxml>=0.8.1",
]

[project.scripts]
mlaude = "mlaude.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mlaude"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
